refac: Command's execution logic
